Adding comments

+++ app/Controller/Admin/ArticleController.php
@@ -2,10 +2,17 @@
 
 namespace App\Controller\Admin;
 
+/**
+     * Controller gérant le rendu des pages liées à l'administration des articles
+*/
+
 class ArticleController extends AppController{
+
+
 	public function __construct(){
 		parent::__construct();
-		$this->loadModel('article');
+		// Chargement des Model propre au controller
+		$this->loadModel('post');
 		$this->loadModel('category');
 		$this->loadModel('comment');
 	}
@@ -14,9 +21,9 @@ class ArticleController extends AppController{
 		$message = '';
 		if(!empty($_POST)){
 			if(isset($_POST['deleteArticle'])){
-				if($this->articleModel->get($_POST['deleteArticle'])){
+				if($this->postModel->get($_POST['deleteArticle'])){
 					$this->commentModel->deleteFromArticle($_POST['deleteArticle']);
-					if($this->articleModel->delete($_POST['deleteArticle'])){
+					if($this->postModel->delete($_POST['deleteArticle'])){
 						$message = ['type' => 'success', 'message' => 'L\'article a bien été supprimé'];
 						
 					}
@@ -24,47 +31,55 @@ class ArticleController extends AppController{
 			}
 			if(isset($_POST['deleteCategory'])){
 				if($this->categoryModel->get($_POST['deleteCategory'])){
-					if($this->categoryModel->delete($_POST['deleteCategory'])){
+					if(empty($this->postModel->getPostsByCategory($_POST['deleteCategory']))){
+						if($this->categoryModel->delete($_POST['deleteCategory'])){
 						$message = ['type' => 'success', 'message' => 'La catégorie a bien été supprimé'];
+						} else {
+							$message = ['type' => 'warning', 'message' => 'Une erreur est survenue'];
+						}
+					} else {
+						$message = ['type' => 'danger', 'message' => 'Vous ne pouvez pas supprimer cette catégorie, des articles y sont associés.'];
+						}
 					}
-				}
 			}
 		}
-		$articles = $this->articleModel->getAll();
-		$lastArticles = $this->articleModel->getLast(5);
+		$posts = $this->postModel->getAll();
+		$lastPosts = $this->postModel->getLast(5);
 		$categories = $this->categoryModel->getAll();
-		$lastComments = $this->commentModel->findLast(10);
-		$commentsByArticle = $this->commentModel->findAllByArticle($articles);
-		$this->render('admin.index', compact('articles', 'categories', 'message', 'lastComments', 'commentsByArticle', 'lastArticles'));
+		$lastComments = $this->commentModel->getLast(10);
+		$reportedComments = $this->commentModel->getReportedComments();
+		$commentsByArticle = $this->commentModel->getAllByPosts($posts);
+		$this->render('admin.index', compact('posts', 'categories', 'message', 'lastComments', 'commentsByArticle', 'lastPosts', 'reportedComments'));
 	}
 
 	public function newArticle(){
-		$categories = $this->categoryModel->extraction('id', 'categorie');
+		$categories = $this->categoryModel->extraction('id', 'title');
 		$form = new \Core\HTML\BootstrapForm();
 
 		if(!empty($_POST)){
-			$titre = (string) $_POST['titre'];
-			$contenu = (string) $_POST['contenu'];
-			$categorie_id = (int) $_POST['categorie_id'];
-			$new = $this->articleModel->new(['titre' => $titre, 'contenu' => $contenu, 'categorie_id' => $categorie_id]);
+			$title = (string) $_POST['title'];
+			$content = (string) $_POST['content'];
+			$category_id = (int) $_POST['category_id'];
+			$new = $this->postModel->add(['title' => $title, 'content' => $content, 'category_id' => $category_id]);
 			if($new){
 				$_SESSION['new'] = true;
-				header('Location: index.php?page=article.edit&id='.$this->app->getDb()->lastInsertId());
+				unset($_SESSION['new']);
+				$message = ['type' => 'success', 'message' => 'L\'article a bien été créé'];
 			}
 		}
-		$this->render('admin.article.create', compact('form', 'categories'));
+		$this->render('admin.article.create', compact('form', 'categories', 'message'));
 	}
 
 	public function editArticle(){
 		$message = '';
-		$categories = $this->categoryModel->extraction('id', 'categorie');
+		$categories = $this->categoryModel->extraction('id', 'title');
 	// On regarde si l'on doit mettre à jour un article avec la présence de données en POST et d'un ID en get
 		if(!empty($_POST) && !empty($_GET)){
-			$titre = (string) $_POST['titre'];
-			$contenu = (string) $_POST['contenu'];
-			$categorie_id = (int) $_POST['categorie_id'];
+			$title = (string) $_POST['title'];
+			$content = (string) $_POST['content'];
+			$category_id = (int) $_POST['category_id'];
 			$articleId = (int) $_GET['id'];
-			$update = $this->articleModel->update($articleId, ['titre' => $titre, 'contenu' => $contenu, 'categorie_id' => $categorie_id]);
+			$update = $this->postModel->update($articleId, ['title' => $title, 'content' => $content, 'category_id' => $category_id]);
 			if($update){
 				$message = ['type' => 'success', 'message' => 'L\'article a bien été mis à jour.'];
 			}
@@ -72,7 +87,7 @@ class ArticleController extends AppController{
 	// On regarde si l'on vient éditer un article avec la présence d'un id
 		if(!empty($_GET)){ 
 			$articleId = (int) $_GET['id'];
-			$article = $this->articleModel->get($articleId);
+			$article = $this->postModel->get($articleId);
 			if(isset($_SESSION['new'])){
 				if($_SESSION['new']){
 					unset($_SESSION['new']);
@@ -81,11 +96,6 @@ class ArticleController extends AppController{
 				}
 			}
 		} else {
-			$message = "
-			<div class='alert alert-warning'>
-				Aucun article sélectionné
-			</div> <!-- div.alert alert-success -->
-			";
 			$message = ['type' => 'warning', 'message' => 'Aucun article sélectionné'];
 		}
 		$form = new \Core\HTML\BootstrapForm($article);
